Changed filename; Merged 2 versions of userRoutes.js & userControllers.js nto single files
